requirements.txt
anthropic>=0.92.0
openai>=1.0.0
langgraph>=0.2.0
python-dotenv>=1.0.0
pymysql>=1.1.0
sqlalchemy>=2.0.0
flask>=3.0.0
requests>=2.31.0
